Removed secondary class, clean up and commenting
Removed BackupJob.cs moving the logic into the main form. This may not be ideal and I may be circling back around to it as I like the idea of being able to plug in either CLI or GUI depending on use / application.
